Restructure tests so that #1442 is easier to resolve in the future
