+++ codegen-client/src/test/kotlin/software/amazon/smithy/rust/codegen/client/smithy/protocols/eventstream/EventStreamBaseRequirements.kt
@@ -1,0 +1,53 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package software.amazon.smithy.rust.codegen.client.smithy.protocols.eventstream
+
+import software.amazon.smithy.model.Model
+import software.amazon.smithy.model.shapes.ServiceShape
+import software.amazon.smithy.model.shapes.ShapeId
+import software.amazon.smithy.model.shapes.StructureShape
+import software.amazon.smithy.rust.codegen.client.smithy.ClientCodegenContext
+import software.amazon.smithy.rust.codegen.client.smithy.customize.CombinedClientCodegenDecorator
+import software.amazon.smithy.rust.codegen.client.testutil.clientTestRustSettings
+import software.amazon.smithy.rust.codegen.client.testutil.testSymbolProvider
+import software.amazon.smithy.rust.codegen.core.rustlang.RustWriter
+import software.amazon.smithy.rust.codegen.core.smithy.CodegenTarget
+import software.amazon.smithy.rust.codegen.core.smithy.RustSymbolProvider
+import software.amazon.smithy.rust.codegen.core.smithy.generators.BuilderGenerator
+import software.amazon.smithy.rust.codegen.core.smithy.generators.implBlock
+import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestRequirements
+
+abstract class EventStreamBaseRequirements : EventStreamTestRequirements<ClientCodegenContext> {
+    override fun createCodegenContext(
+        model: Model,
+        symbolProvider: RustSymbolProvider,
+        serviceShape: ServiceShape,
+        protocolShapeId: ShapeId,
+        codegenTarget: CodegenTarget,
+    ): ClientCodegenContext = ClientCodegenContext(
+        model,
+        symbolProvider,
+        serviceShape,
+        protocolShapeId,
+        clientTestRustSettings(),
+        CombinedClientCodegenDecorator(emptyList()),
+    )
+
+    override fun createSymbolProvider(model: Model): RustSymbolProvider = testSymbolProvider(model)
+
+    override fun renderBuilderForShape(
+        writer: RustWriter,
+        codegenContext: ClientCodegenContext,
+        shape: StructureShape,
+    ) {
+        BuilderGenerator(codegenContext.model, codegenContext.symbolProvider, shape).apply {
+            render(writer)
+            writer.implBlock(shape, codegenContext.symbolProvider) {
+                renderConvenienceMethod(writer)
+            }
+        }
+    }
+}
