+++ codegen-client/src/test/kotlin/software/amazon/smithy/rust/codegen/client/smithy/protocols/eventstream/EventStreamMarshallerGeneratorTest.kt
@@ -3,27 +3,19 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.rust.codegen.client.smithy.protocols.serialize
+package software.amazon.smithy.rust.codegen.client.smithy.protocols.eventstream
 
 import org.junit.jupiter.api.extension.ExtensionContext
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
-import software.amazon.smithy.model.Model
-import software.amazon.smithy.model.shapes.ServiceShape
-import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.rust.codegen.client.smithy.ClientCodegenContext
-import software.amazon.smithy.rust.codegen.client.smithy.customize.CombinedClientCodegenDecorator
-import software.amazon.smithy.rust.codegen.client.testutil.clientTestRustSettings
-import software.amazon.smithy.rust.codegen.client.testutil.testSymbolProvider
 import software.amazon.smithy.rust.codegen.core.smithy.CodegenTarget
 import software.amazon.smithy.rust.codegen.core.smithy.RuntimeType
-import software.amazon.smithy.rust.codegen.core.smithy.RustSymbolProvider
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.Protocol
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.serialize.EventStreamMarshallerGenerator
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestModels
-import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestRequirements
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestTools
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestVariety
 import software.amazon.smithy.rust.codegen.core.testutil.TestEventStreamProject
@@ -44,24 +36,7 @@ class EventStreamMarshallerGeneratorTest {
     fun test(testCase: EventStreamTestModels.TestCase) {
         EventStreamTestTools.runTestCase(
             testCase,
-            object : EventStreamTestRequirements<ClientCodegenContext> {
-                override fun createCodegenContext(
-                    model: Model,
-                    symbolProvider: RustSymbolProvider,
-                    serviceShape: ServiceShape,
-                    protocolShapeId: ShapeId,
-                    codegenTarget: CodegenTarget,
-                ): ClientCodegenContext = ClientCodegenContext(
-                    model,
-                    symbolProvider,
-                    serviceShape,
-                    protocolShapeId,
-                    clientTestRustSettings(),
-                    CombinedClientCodegenDecorator(emptyList()),
-                )
-
-                override fun createSymbolProvider(model: Model): RustSymbolProvider = testSymbolProvider(model)
-
+            object : EventStreamBaseRequirements() {
                 override fun renderGenerator(
                     codegenContext: ClientCodegenContext,
                     project: TestEventStreamProject,
