+++ codegen-server/src/test/kotlin/software/amazon/smithy/rust/codegen/server/smithy/protocols/eventstream/EventStreamMarshallerGeneratorTest.kt
@@ -3,33 +3,27 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package software.amazon.smithy.rust.codegen.server.smithy.protocols.serialize
+package software.amazon.smithy.rust.codegen.server.smithy.protocols.eventstream
 
 import org.junit.jupiter.api.extension.ExtensionContext
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.ArgumentsProvider
 import org.junit.jupiter.params.provider.ArgumentsSource
-import software.amazon.smithy.model.Model
-import software.amazon.smithy.model.shapes.ServiceShape
-import software.amazon.smithy.model.shapes.ShapeId
+import software.amazon.smithy.model.shapes.StructureShape
+import software.amazon.smithy.rust.codegen.core.rustlang.RustWriter
 import software.amazon.smithy.rust.codegen.core.smithy.CodegenTarget
 import software.amazon.smithy.rust.codegen.core.smithy.RuntimeType
-import software.amazon.smithy.rust.codegen.core.smithy.RustSymbolProvider
+import software.amazon.smithy.rust.codegen.core.smithy.generators.implBlock
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.Protocol
 import software.amazon.smithy.rust.codegen.core.smithy.protocols.serialize.EventStreamMarshallerGenerator
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestModels
-import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestRequirements
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestTools
 import software.amazon.smithy.rust.codegen.core.testutil.EventStreamTestVariety
 import software.amazon.smithy.rust.codegen.core.testutil.TestEventStreamProject
 import software.amazon.smithy.rust.codegen.core.testutil.TestRuntimeConfig
-import software.amazon.smithy.rust.codegen.server.smithy.RustCodegenServerPlugin
 import software.amazon.smithy.rust.codegen.server.smithy.ServerCodegenContext
-import software.amazon.smithy.rust.codegen.server.smithy.ServerSymbolProviders
-import software.amazon.smithy.rust.codegen.server.smithy.testutil.ServerTestSymbolVisitorConfig
-import software.amazon.smithy.rust.codegen.server.smithy.testutil.serverTestRustSettings
-import software.amazon.smithy.rust.codegen.server.smithy.testutil.serverTestSymbolProvider
+import software.amazon.smithy.rust.codegen.server.smithy.generators.ServerBuilderGenerator
 import java.util.stream.Stream
 
 class MarshallTestCasesProvider : ArgumentsProvider {
@@ -46,38 +40,7 @@ class EventStreamMarshallerGeneratorTest {
     fun test(testCase: EventStreamTestModels.TestCase) {
         EventStreamTestTools.runTestCase(
             testCase,
-            object : EventStreamTestRequirements<ServerCodegenContext> {
-                override fun createCodegenContext(
-                    model: Model,
-                    symbolProvider: RustSymbolProvider,
-                    serviceShape: ServiceShape,
-                    protocolShapeId: ShapeId,
-                    codegenTarget: CodegenTarget,
-                ): ServerCodegenContext {
-                    val settings = serverTestRustSettings()
-                    val serverSymbolProviders = ServerSymbolProviders.from(
-                        model,
-                        serviceShape,
-                        ServerTestSymbolVisitorConfig,
-                        settings.codegenConfig.publicConstrainedTypes,
-                        RustCodegenServerPlugin::baseSymbolProvider,
-                    )
-                    return ServerCodegenContext(
-                        model,
-                        symbolProvider,
-                        serviceShape,
-                        protocolShapeId,
-                        settings,
-                        serverSymbolProviders.unconstrainedShapeSymbolProvider,
-                        serverSymbolProviders.constrainedShapeSymbolProvider,
-                        serverSymbolProviders.constraintViolationSymbolProvider,
-                        serverSymbolProviders.pubCrateConstrainedShapeSymbolProvider,
-                    )
-                }
-
-                override fun createSymbolProvider(model: Model): RustSymbolProvider =
-                    serverTestSymbolProvider(model)
-
+            object : EventStreamBaseRequirements() {
                 override fun renderGenerator(
                     codegenContext: ServerCodegenContext,
                     project: TestEventStreamProject,
@@ -92,6 +55,19 @@ class EventStreamMarshallerGeneratorTest {
                         protocol.structuredDataSerializer(project.operationShape),
                         testCase.requestContentType,
                     ).render()
+                }
+
+                override fun renderBuilderForShape(
+                    writer: RustWriter,
+                    codegenContext: ServerCodegenContext,
+                    shape: StructureShape,
+                ) {
+                    ServerBuilderGenerator(codegenContext, shape).apply {
+                        render(writer)
+                        writer.implBlock(shape, codegenContext.symbolProvider) {
+                            renderConvenienceMethod(writer)
+                        }
+                    }
                 }
             },
             CodegenTarget.SERVER,
